requirements-django.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Pillow==10.3.0
python-multipart==0.0.12
sendgrid==6.11.0
pyjwt==2.9.0
python-dateutil==2.9.0
